Respond to the events being sent from the SyncManager to update the GUI.
* Tomboy/SyncDialog.cs: Respond to the events being sent from the
  SyncManager to update the GUI.
* Tomboy/SyncManager.cs: Do the synchronization on a thread so the GUI
  doesn't lock.

svn path=/branches/tomboy-sync/; revision=1238

+++ Tomboy/SyncManager.cs
@@ -3,14 +3,111 @@ using System.IO;
 using System.Xml;
 using System.Collections.Generic;
 
+using System.Threading;
+
 using Mono.Unix;
 
 namespace Tomboy
 {
+	public enum SyncState {
+		/// <summary>
+		/// The synchronization thread is not running
+		/// </summary>
+		Idle,
+		
+		/// <summary>
+		/// Connecting to the server
+		/// </summary>
+		Connecting,
+		
+		/// <summary>
+		/// Acquiring the right to be the exclusive sync client
+		/// </summary>
+		AcquiringLock,
+
+		/// <summary>
+		/// Another client is currently synchronizing
+		/// </summary>
+		Locked,
+		
+		/// <summary>
+		/// Preparing to download new/updated notes from the server.  This also
+		/// includes checking for note title name conflicts.
+		/// </summary>
+		PrepareDownload,
+
+		/// <summary>
+		/// Downloading notes from the server
+		/// </summary>
+		Downloading,
+		
+		/// <summary>
+		/// Checking for files to send to the server
+		/// </summary>
+		PrepareUpload,
+
+		/// <summary>
+		/// Uploading new/changed notes from the client
+		/// </summary>
+		Uploading,
+		
+		/// <summary>
+		/// Deleting notes from the server
+		/// </summary>
+		DeleteServerNotes,
+		
+		/// <summary>
+		/// Committing Changes to the server
+		/// </summary>
+		CommittingChanges,
+		
+		/// <summary>
+		/// SyncSuccess
+		/// </summary>
+		Succeeded,
+		
+		/// <summary>
+		/// The synchronization failed
+		/// </summary>
+		Failed
+	};
+	
+	public enum NoteSyncType {
+		UploadNew,
+		UploadModified,
+		DownloadNew,
+		DownloadModified,
+		DeleteFromServer,
+		DeleteFromClient
+	};
+	
+	/// <summary>
+	/// Handle state SyncManager state changes
+	/// </summary>
+	public delegate void SyncStateChangedHandler (SyncState state);
+	
+	/// <summary>
+	/// Handle when notes are uploaded, downloaded, or deleted
+	/// </summary>
+	public delegate void NoteSyncHandler (string noteTitle, NoteSyncType type); 
+	
 	public class SyncManager
 	{
 		//private static SyncServer server;
 		private static SyncClient client;
+		private static SyncState state = SyncState.Idle;
+		
+		private static Thread syncThread = null;
+		
+		/// <summary>
+		/// Emitted when the state of the synchronization changes
+		/// </summary>
+		public static event SyncStateChangedHandler StateChanged;
+		
+		/// <summary>
+		/// Emmitted when a file is uploaded, downloaded, or deleted.
+		/// </summary>
+		public static event NoteSyncHandler NoteSynchronized;
 		
 		static SyncManager ()
 		{
@@ -19,33 +116,55 @@ namespace Tomboy
 		}
 		
 		public static void PerformSynchronization ()
+		{
+			if (syncThread != null) {
+				// A synchronization thread is already running
+				Tomboy.SyncDialog.Present ();
+				return;
+			}
+			
+			syncThread = new Thread (new ThreadStart (SynchronizationThread));
+			syncThread.IsBackground = true;
+			syncThread.Start ();
+		}
+		
+		/// <summary>
+		/// The function that does all of the work
+		/// </summary>
+		public static void SynchronizationThread ()
 		{			
 			SyncServer server;
+			SetState (SyncState.Connecting);
 			try {
 				server = new FileSystemSyncServer ();
 			} catch (Exception e) {
+				SetState (SyncState.Failed);
 				Logger.Log ("Exception while creating SyncServer: {0}\n{1}", e.Message, e.StackTrace);
+				SetState (SyncState.Idle);
 				return;
-				// TODO: This should be a GUI dialog
+				// TODO: Figure out a clever way to get the specific error up to the GUI
 			}
 			
-			SyncDialog syncDialog = Tomboy.SyncDialog;
-			syncDialog.ProgressText = "Contacting Server...";
-			syncDialog.Response += OnSyncDialogResponse;
-			syncDialog.Show ();
+//			SyncDialog syncDialog = Tomboy.SyncDialog;
+//			syncDialog.ProgressText = "Contacting Server...";
+//			syncDialog.Response += OnSyncDialogResponse;
+//			syncDialog.Show ();
 			
-			// TODO: How to actually do transactions?  Is it even possible?
+			SetState (SyncState.AcquiringLock);
+			// TODO: We should really throw exceptions from BeginSyncTransaction ()
 			if (!server.BeginSyncTransaction ()) {
+				SetState (SyncState.Locked);
 				Logger.Log ("PerformSynchronization: Server locked, try again later");
-				syncDialog.ProgressText = "Sync failed: Server locked, try again later";
-				syncDialog.CloseSensitive = true;
+//				syncDialog.ProgressText = "Sync failed: Server locked, try again later";
+//				syncDialog.CloseSensitive = true;
 				return;
 			}
 			int latestServerRevision = server.LatestRevision;
 			int newRevision = latestServerRevision + 1;
 			
-			syncDialog.ProgressText = "Getting note updates...";
-			syncDialog.ProgressFraction = 0.1;
+			SetState (SyncState.PrepareDownload);
+//			syncDialog.ProgressText = "Getting note updates...";
+//			syncDialog.ProgressFraction = 0.1;
 			
 			// Handle notes modified or added on server
 			Logger.Debug ("Sync: GetNoteUpdatesSince rev " + client.LastSynchronizedRevision.ToString ());
@@ -104,8 +223,11 @@ namespace Tomboy
 			
 			double updateProgressInterval = 0.4 / (double)noteUpdates.Count;
 			
+			if (noteUpdates.Count > 0)
+				SetState (SyncState.Downloading);
+
 			foreach (NoteUpdate note in noteUpdates.Values) {
-				syncDialog.ProgressFraction += updateProgressInterval;
+//				syncDialog.ProgressFraction += updateProgressInterval;
 				Note existingNote = FindNoteByUUID (note.UUID);
 				
 				if (existingNote == null) {// TODO: not so simple...what if I deleted the note?
@@ -114,45 +236,65 @@ namespace Tomboy
 
 					existingNote = NoteMgr.CreateWithGuid (note.Title, note.UUID);
 					existingNote.LoadForeignNoteXml (note.XmlContent);					
-					syncDialog.AddUpdateItem (existingNote.Title, Catalog.GetString ("Downloading new note"));
+					if (NoteSynchronized != null)
+						NoteSynchronized (existingNote.Title, NoteSyncType.DownloadNew);
+//					syncDialog.AddUpdateItem (existingNote.Title, Catalog.GetString ("Downloading new note"));
 				} else if (note.XmlContent == string.Empty &&
 				         existingNote.ChangeDate.CompareTo (client.LastSyncDate) <= 0) {
-					syncDialog.AddUpdateItem (existingNote.Title, Catalog.GetString ("Deleting note on client"));
+						if (NoteSynchronized != null)
+							NoteSynchronized (existingNote.Title, NoteSyncType.DeleteFromClient);
+//					syncDialog.AddUpdateItem (existingNote.Title, Catalog.GetString ("Deleting note on client"));
 					NoteMgr.Delete (existingNote);
 				} else if (existingNote.ChangeDate.CompareTo (client.LastSyncDate) <= 0) {
 					existingNote.LoadForeignNoteXml (note.XmlContent);// TODO: live update of note XML
-					syncDialog.AddUpdateItem (existingNote.Title, Catalog.GetString ("Downloading update"));
+					if (NoteSynchronized != null)
+						NoteSynchronized (existingNote.Title, NoteSyncType.DownloadModified);
+//					syncDialog.AddUpdateItem (existingNote.Title, Catalog.GetString ("Downloading update"));
 				} else {
 					// TODO: handle conflicts
 				}
 			}
 			
 			// TODO: Add following updates to syncDialog treeview
-			syncDialog.ProgressText = "Sending note updates...";
-			syncDialog.ProgressFraction = 0.5;
+//			syncDialog.ProgressText = "Sending note updates...";
+//			syncDialog.ProgressFraction = 0.5;
 			
-			// Upload notes modified or added on client
-			List<Note> newOrModifiedNotes = new List<Note> ();			
+			SetState (SyncState.PrepareUpload);
+			// Look through all the notes modified on the client
+			// and upload new or modified ones to the server
+			List<Note> newOrModifiedNotes = new List<Note> ();
+Logger.Debug ("SYNC: client.LastSyncDate = {0}", client.LastSyncDate.ToString ());
 			foreach (Note note in NoteMgr.Notes) {
+if (note.Title.CompareTo ("Start Here") == 0) {
+	Logger.Debug ("SYNC: Start Here Revision = {0}", note.Revision);
+	Logger.Debug ("SYNC: Start Here Change Date = {0}", note.ChangeDate.ToString ());
+}
 				if (note.Revision == -1) {
 					// This is a new note that has never been synchronized to the server
 					newOrModifiedNotes.Add (note);
-					syncDialog.AddUpdateItem (note.Title, Catalog.GetString ("Adding new file to server"));
-				} else if (note.Revision < client.LastSynchronizedRevision &&
-				    	note.ChangeDate.CompareTo (client.LastSyncDate) > 0) {
+					if (NoteSynchronized != null)
+						NoteSynchronized (note.Title, NoteSyncType.UploadNew);
+//					syncDialog.AddUpdateItem (note.Title, Catalog.GetString ("Adding new file to server"));
+				} else if (note.Revision <= client.LastSynchronizedRevision &&
+				    	note.ChangeDate > client.LastSyncDate) {
 					newOrModifiedNotes.Add (note);
-					syncDialog.AddUpdateItem (note.Title, Catalog.GetString ("Uploading changes"));
+					if (NoteSynchronized != null)
+						NoteSynchronized (note.Title, NoteSyncType.UploadModified);
+//					syncDialog.AddUpdateItem (note.Title, Catalog.GetString ("Uploading changes"));
 				}
 			}
 			// Apply this revision number to all new/modified notes since last sync
 			// TODO: Should revision info be stored in note, or a seperate file?
+			// TODO: This may actually cause future problems if our network connection
+			// dies after changing the revision but never really gets uploaded to the
+			// server.
 			foreach (Note note in newOrModifiedNotes) {
 				note.Revision = newRevision;
 				note.Save ();
 			}
 			Logger.Debug ("Sync: Uploading " + newOrModifiedNotes.Count.ToString () + " note updates");
+			SetState (SyncState.Uploading);
 			server.UploadNotes (newOrModifiedNotes);
-			
 			
 			// Handle notes deleted on client
 			List<string> locallyDeletedUUIDs = new List<string> ();			
@@ -160,21 +302,29 @@ namespace Tomboy
 				if (FindNoteByUUID (noteUUID) == null) {
 					locallyDeletedUUIDs.Add (noteUUID);
 					// TODO: Cough, ugh.  Can we get the title at this point?
-					syncDialog.AddUpdateItem (noteUUID, Catalog.GetString ("Deleting from server"));
+					if (NoteSynchronized != null)
+						NoteSynchronized ("FIXME: Get the deleted note's title!", NoteSyncType.DeleteFromServer);
+//					syncDialog.AddUpdateItem (noteUUID, Catalog.GetString ("Deleting from server"));
 				}
-			}			
-			server.DeleteNotes (locallyDeletedUUIDs);
+			}
+			if (locallyDeletedUUIDs.Count > 0) {
+				SetState (SyncState.DeleteServerNotes);
+				server.DeleteNotes (locallyDeletedUUIDs);
+			}
 			
+			SetState (SyncState.CommittingChanges);
 			bool commitResult = server.CommitSyncTransaction ();
 			if (commitResult) {
-				syncDialog.ProgressFraction = 1.0;
-				syncDialog.ProgressText = "Sync completed successfully!";
+//				syncDialog.ProgressFraction = 1.0;
+//				syncDialog.ProgressText = "Sync completed successfully!";
+				SetState (SyncState.Succeeded);
+			} else {
+//				syncDialog.ProgressFraction = 0;
+//				syncDialog.ProgressText = "Sync failed!";
+				SetState (SyncState.Failed);
+				// TODO: Figure out a way to let the GUI know what exactly failed
 			}
-			else {
-				syncDialog.ProgressFraction = 0;
-				syncDialog.ProgressText = "Sync failed!";
-			}
-			syncDialog.CloseSensitive = true;
+//			syncDialog.CloseSensitive = true;
 			
 			// This should be equivalent to newRevision
 			client.LastSynchronizedRevision = server.LatestRevision;
@@ -182,6 +332,9 @@ namespace Tomboy
 			client.LastSyncDate = DateTime.Now;		
 			
 			Logger.Debug ("Sync: New revision: {0}", client.LastSynchronizedRevision);
+			
+			SetState (SyncState.Idle);
+			syncThread = null;
 		}
 		
 		private static Note FindNoteByUUID (string uuid)
@@ -194,12 +347,35 @@ namespace Tomboy
 			get { return Tomboy.DefaultNoteManager; }
 		}
 		
-		private static void OnSyncDialogResponse (object sender, Gtk.ResponseArgs args)
+//		private static void OnSyncDialogResponse (object sender, Gtk.ResponseArgs args)
+//		{
+//			SyncDialog dialog = sender as SyncDialog;
+//			dialog.Hide ();
+//			dialog.Destroy ();
+//		}
+		
+		#region Public Properties
+		/// <summary>
+		/// The state of the SyncManager (lame comment, duh!)
+		/// </summary>
+		public SyncState State
 		{
-			SyncDialog dialog = sender as SyncDialog;
-			dialog.Hide ();
-			dialog.Destroy ();
+			get { return state; }
 		}
+		#endregion // Public Properties
+		
+		#region Private Methods
+		private static void SetState (SyncState newState)
+		{
+			state = newState;
+			if (StateChanged != null) {
+				// Notify the event handlers
+				try {
+					StateChanged (state);
+				} catch {}
+			}
+		}
+		#endregion // Private Methods
 	}
 	
 	public class NoteUpdate
